requirements.txt
langchain-community
openai
youtube-transcript-api
tiktoken
pypdf